+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Unreleased
 
+* Fix deprecation on password ->isCorrect() when hash is null - e.g. if a user's password has never been initialised.
+
 ### v1.2.1 (2022-10-24)
 
 * Checking MX records for valid emails is unreliable and hes been dropped from symfony/validation. Utilise 'strict' mode instead to check the address looks valid.